modified build script

+++ build.sh
@@ -1,4 +1,4 @@
-
+#!/usr/bin/env bash
 cmake -G Ninja -B build \
   -DCMAKE_C_COMPILER=clang -DCMAKE_CXX_COMPILER=clang++ \
   -DCMAKE_BUILD_TYPE=RelWithDebInfo \
@@ -6,5 +6,8 @@ cmake -G Ninja -B build \
 cmake --build build/
 if [[ -e "build/bin/resources" ]]; then
   ln -sr resources build/bin/resources
+fi
+if [[ -e app_local_ ]]; then
+  exit
 fi
 ln -s build/bin/application app_local_
